Fix warnings like missing return values or signed unsigned conflicts

+++ src/rasta/c/rastadeferqueue.c
@@ -10,7 +10,7 @@
  * @return -1 if there is no element with the specified @p seq_nr, index of the element otherwise
  */
 int find_index(struct defer_queue * queue, unsigned long seq_nr){
-    int index = 0;
+    uint32_t index = 0;
     // Check is now required because queue stored pointers!
     if(!queue->elements[index].packet)
         return -1;
@@ -20,7 +20,7 @@ int find_index(struct defer_queue * queue, unsigned long seq_nr){
     uint32_t maxCount = queue->max_count;
     while ( index < maxCount && sequeceNrCMP != seq_nr ) ++index;
 
-    return ( index == queue->max_count ? -1 : index );
+    return ( index == queue->max_count ? -1 : (int)index );
 }
 
 int cmpfkt(const void * a, const void * b){
@@ -105,7 +105,7 @@ void deferqueue_remove(struct defer_queue * queue, unsigned long seq_nr){
         return;
     }
 
-    if(index != (queue->count - 1)){
+    if(index != (int)(queue->count - 1)){
         // element to delete isn't at the last position
         // to be able to add the next element to the last position without overriding something
         // the currently last element is moved to the index where the element to delete is located
@@ -156,10 +156,10 @@ int deferqueue_smallest_seqnr(struct defer_queue * queue){
     unsigned long smallest = 0xFFFFFFFF;
 
     // naive implementation of search. performance shouldn't be an issue as the amount of messages in the queue is small
-    for (int i = 0; i < queue->max_count; ++i) {
+    for (unsigned int i = 0; i < queue->max_count; ++i) {
         if(queue->elements[i].packet->sequence_number < smallest){
             smallest = queue->elements[i].packet->sequence_number;
-            index = i;
+            index = (int)i;
         }
     }
